requirements_new.txt
# Core dependencies
playwright>=1.40.0
pyyaml>=6.0
reportlab>=4.0.0
matplotlib>=3.7.0
pandas>=2.0.0

# Optional dependencies for enhanced features
# selenium>=4.15.0  # Alternative browser automation
# requests>=2.31.0  # HTTP client for API testing
# beautifulsoup4>=4.12.0  # HTML parsing
# pillow>=10.0.0  # Image processing
# jinja2>=3.1.0  # Template engine for reports
